backend/requirements.txt
fastapi 
uvicorn 
groq 

gtts 
python-multipart 
python-dotenv